requirements.txt
PyGithub
loguru
wrapt>=1.16.0
